backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.34
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
cryptography==43.0.1
redis==5.0.8
celery==5.4.0
httpx==0.27.2
alembic
google-api-python-client==2.148.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
python-multipart==0.0.9
email-validator>=2.0.0
